requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-environ==0.11.2
mysqlclient==2.2.5
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
